+++ CHANGES.md
@@ -11,3 +11,39 @@ none
 #### Features
 
 - Components [tree] Add automatic folding/expand judgment, With prop nodeContentClickAction: TreeNodeContentClickActionEnum.CHECK.
+
+### 0.1.2
+
+_2024-03-23_
+
+#### Bug fixes
+
+none
+
+#### Features
+
+- Components [tree] remove "node-content" slot,add default slot for node content.
+
+### 0.1.3
+
+_2024-03-23_
+
+#### Bug fixes
+
+none
+
+#### Features
+
+- Components [tree] remove "node-content" slot,add default slot for node content.
+
+### 0.1.4
+
+_2024-03-25_
+
+#### Bug fixes
+
+fix display error!
+
+#### Features
+
+none